몬스터 웨이브

+++ GameEngineBase/JPSHeap.h
@@ -22,11 +22,11 @@ public:
 
 	float DistanceSqrt(const JPSCoord& rhs)
 	{
-		Int32 absX = std::abs(m_x - rhs.m_x);
-		Int32 absY = std::abs(m_y - rhs.m_y);
+		Int32 absX = static_cast<Int32>(std::abs(m_x - rhs.m_x));
+		Int32 absY = static_cast<Int32>(std::abs(m_y - rhs.m_y));
 
-		Int32 diagDist = std::fmin(absX, absY);
-		Int32 straightDist = std::fmax(absX, absY) - diagDist;
+		Int32 diagDist = static_cast<Int32>(std::fmin(absX, absY));
+		Int32 straightDist = static_cast<Int32>(std::fmax(absX, absY)) - diagDist;
 		return diagDist * 1.414213562373095f + straightDist;
 	}
 };
